requirements.txt
celery==4.3.0
Django==2.2.3
django-controlcenter==0.2.9
django-import-export==1.2.0
django-localflavor==2.2
 -e git://github.com/darklow/django-suit.git@v2#egg=django-suit
redis==3.2.1